Revert group errors extraction
