+++ packages/cli/test/workspace/errors.test.ts
@@ -16,7 +16,7 @@
 import _ from 'lodash'
 import { errors as wsErrors } from '@salto-io/workspace'
 import { SaltoError, ElemID } from '@salto-io/adapter-api'
-import { groupRelatedErrors } from '../../src/workspace/errors'
+import { groupRelatedErrors, UnresolvedReferenceGroupError } from '../../src/workspace/errors'
 
 describe('groupRelatedErrors', () => {
   let nonGroupedError: SaltoError
@@ -49,6 +49,56 @@ describe('groupRelatedErrors', () => {
       })
       it('should keep non grouped errors unchanged', () => {
         expect(groupedErrors).toContain(nonGroupedError)
+      })
+      describe('UnresolvedReferenceGroupError', () => {
+        let groupError: UnresolvedReferenceGroupError
+        beforeEach(() => {
+          groupError = groupedErrors.find(
+            err => err instanceof UnresolvedReferenceGroupError
+          ) as UnresolvedReferenceGroupError
+        })
+        it('should be in the grouped errors', () => {
+          expect(groupError).toBeDefined()
+        })
+        it('should contain the number of errors it grouped', () => {
+          expect(groupError.message).toContain(numSources.toString())
+        })
+        it('should contain the target', () => {
+          expect(groupError.message).toContain(target.getFullName())
+        })
+      })
+    })
+
+    describe('with references to targets nested in other targets', () => {
+      let baseTarget: ElemID
+      beforeEach(() => {
+        baseTarget = new ElemID('test', 'target', 'field', 'bla')
+        originalErrors = sources.map(
+          (elemID, idx) => new wsErrors.UnresolvedReferenceValidationError(
+            { elemID, target: baseTarget.createNestedID(..._.times(idx, () => 'a')) }
+          )
+        )
+        groupedErrors = groupRelatedErrors(originalErrors)
+      })
+      it('should group to the outermost target', () => {
+        expect(groupedErrors).toHaveLength(1)
+        const [error] = groupedErrors
+        expect(error).toBeInstanceOf(UnresolvedReferenceGroupError)
+        expect(error.message).toContain(baseTarget.getFullName())
+      })
+    })
+
+    describe('with a single unresolved reference', () => {
+      let target: ElemID
+      beforeEach(() => {
+        target = new ElemID('test', 'target')
+        originalErrors = [
+          new wsErrors.UnresolvedReferenceValidationError({ elemID: sources[0], target }),
+        ]
+        groupedErrors = groupRelatedErrors(originalErrors)
+      })
+      it('should not create a group error', () => {
+        expect(originalErrors).toEqual(groupedErrors)
       })
     })
   })
